agent/pyproject.toml
[project]
name = "basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
